requirements.txt
Flask
langchain
openai
pinecone-client
pymongo
pymodm
PyPDF2
pytest
tiktoken
flask_restx